+++ demo/mkinit_demo_pkg/submod.py
@@ -1,0 +1,4 @@
+# --- submod.py ---
+
+def submod_func():
+    print('This is a submod func in {}'.format(__file__))